+++ library.properties
@@ -1,9 +1,9 @@
-name=OLED MENU
+name=Watch Menu
 version=1.0.0
 author=BAGALOOZY
 maintainer=Bagaloozy <winneymj@gmail.com>
-sentence=OLED Menu system!
-paragraph=OLED Menu system!
+sentence=Watch Menu system!
+paragraph=Watch Menu system!
 category=Display
-url=https://github.com/winneymj/OLED_Menu.git
+url=https://github.com/winneymj/Watch_Menu.git
 architectures=*
